% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dollo.10.48.2.f0.gl0}
\alias{dollo.10.48.2.f0.gl0}
\title{Simulated data from Generalized Dollo model, with seed 2}
\format{A list of 8 variables:
\itemize{
  \item mutCounts: An integer matrix, with columns corresponding to the genomic loci. First and second rows contain the reference and variant alelle counts respectively.
  \item psi: The copy number at this genomic locus, encoded as a list, each element of which is a list of 3, \code{gV}, \code{gN}, and \code{gR},
   corresponding to variant Copy Number (CN), reference genotype CN, and normal CN.
   \code{gV} is a string in an alphabet of {A,B}, where the number of {A}-s denotes the major and the number of {B}-s denotes the minor CN respectively.
  \item filteredMutMatrix: A binary genotype-locus matrix, with element [i,j] is one if genotype i is mutated at locus j and is zero otherwise.
  \item mutPrevalence: For internal use of the package.
  \item missingIndx: For internal use of the package.
  \item phi: For internal use of the package.
  \item trueMutMatrix: For internal use of the package.
  \item psis.single.cell: For internal use of the package.
  \item dolloDirName: For internal use fo the package.
 }}
\usage{
dollo.10.48.2.f0.gl0
}
\description{
A dataset containing the allele counts and genotypes from the Generalized Dollo model.
See (Salehi, Sohrab, Adi Steif, Roth Andrew, Samuel Aparicio, Alexandre Bouchard, and Sohrab P. Shah. “Joint Statistical Inference of Clonal Populations ￼from Single Cell and Bulk Tumour Sequencing Data.” (submitted).)
}
\keyword{datasets}

