% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{estimatePointClustering}
\alias{estimatePointClustering}
\title{Computes a point estimate (using MAXPEAR method) for clustering of the mutations from the results of a ddClone analysis run}
\usage{
estimatePointClustering(expPath, MCMCOptions = list(thinning = 10, burnIn =
  100))
}
\arguments{
\item{expPath}{path to the directory where the ddClone results are stored}

\item{MCMCOptions}{a listing containing MCMC options including \code{thinning} and \code{burnIn} to be used.}
}
\value{
A matrix each row of which is a point estimate of clustering assignment for each mutation. See mcclust::maxpear for more details.
}
\description{
Computes a point estimate (using MAXPEAR method) for clustering of the mutations from the results of a ddClone analysis run
}

