% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{estimatePointPhi}
\alias{estimatePointPhi}
\title{Computes a Monte Carlo estimate for the cellular prevalences of each mutation}
\usage{
estimatePointPhi(expPath, MCMCOptions = list(thinning = 10, burnIn = 100))
}
\arguments{
\item{expPath}{path to the directory where the ddClone results are stored}

\item{MCMCOptions}{a listing containing MCMC options including \code{thinning} and \code{burnIn} to be used.}
}
\value{
a vector containingcontaining the Monte Carlo estimate for each mutation
}
\description{
Computes a Monte Carlo estimate for the cellular prevalences of each mutation
}

