% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{evaluate.clustering}
\alias{evaluate.clustering}
\title{Computes the V-measure for clustering assessment.}
\usage{
evaluate.clustering(labels_true, labels_pred)
}
\arguments{
\item{labels_true}{vector of true clustering assignments where \code{labels_true[[i]]} is the clustering assignment of datapoint \code{i}}

\item{labels_pred}{same as above, except for predicted clustering assignments}
}
\value{
A list the containing the \code{v_measure_score}, and its components, \code{homogeneity}, and \code{completeness}.
}
\description{
See Rosenberg A, Hirschberg J. V-Measure: A Conditional Entropy-Based External Cluster Evaluation Measure. InEMNLP-CoNLL 2007 Jun 28 (Vol. 7, pp. 410-420).
Based on implentation in \url{http://pydoc.net/Python/scikit-learn/0.14.1/sklearn.metrics.cluster.supervised/}
}

