% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make.ddclone.input}
\alias{make.ddclone.input}
\title{Generate a ddClone input object from bulk allele counts and single cell data and saves it in \code{bulkDataPath}.}
\usage{
make.ddclone.input(bulkDataPath, genotypeMatrixPath, outputPath, nameTag = "")
}
\arguments{
\item{bulkDataPath}{Path to a .tsv file that contains allele counts and copy number data. Expects colnames to be "mutation_id", "ref_counts", "var_counts", "normal_cn", "minor_cn", and "major_cn". "minor_cn" and "major_cn" should be integer values for the minimum and maxmum estimated copy number at that locus respectively.}

\item{genotypeMatrixPath}{Path to a .tsv file of binary entries where rows correspond to genotypes and coloumns to mutation IDs respectively.}

\item{outputPath}{What directory should the output be saved into.}

\item{nameTag}{An optional string to be appended to the name of the output object}
}
\value{
A list with appropriate members that could be given as input for ddClone analysis, for instance \code{ddclone::ddclone}
}
\description{
Generate a ddClone input object from bulk allele counts and single cell data and saves it in \code{bulkDataPath}.
}

