% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{relabel.clusters}
\alias{relabel.clusters}
\title{Computes human friendly labels given a numeric vector \code{numericVector}.}
\usage{
relabel.clusters(numericVector)
}
\arguments{
\item{numericVector}{vector of numeric values, could be cluster labels estimated by ddclone.}
}
\value{
A list the same length as \code{numericVector}, preserving cluster assignment, while renaming labels to consecutive numbers.
}
\description{
Computes human friendly labels given a numeric vector \code{numericVector}.
}

