% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnbc.xenograft.SA494.T}
\alias{tnbc.xenograft.SA494.T}
\title{my Data}
\format{An object of class \code{list} of length 5.}
\usage{
tnbc.xenograft.SA494.T
}
\description{
A dataset containing the allele counts and genotypes from a TNBC sample
 The variables are as follows:
}
\keyword{datasets}

