% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tnbc.xenograft.SA501.X2}
\alias{tnbc.xenograft.SA501.X2}
\title{Tripple Negative breast cancer Xenograft, sample SA501, timepoint X2}
\format{A list of 5 variables:
\itemize{
  \item mutCounts: An integer matrix, with columns corresponding to the genomic loci. First and second rows contain the reference and variant alelle counts respectively.
  \item psi: The copy number at this genomic locus, encoded as a list, each element of which is a list of 3, \code{gV}, \code{gN}, and \code{gR},
   corresponding to variant Copy Number (CN), reference genotype CN, and normal CN.
   \code{gV} is a string in an alphabet of {A,B}, where the number of {A}-s denotes the major and the number of {B}-s denotes the minor CN respectively.
  \item filteredMutMatrix: A binary genotype-locus matrix, with element [i,j] is one if genotype i is mutated at locus j and is zero otherwise.
  \item mutPrevalence: For internal use of the package.
  \item missingIndx: For internal use of the package.
 }}
\usage{
tnbc.xenograft.SA501.X2
}
\description{
A dataset containing the allele counts and genotypes from a Tripple negative breast cancer study (see
\cite{Eirew P, Steif A, Khattra J, Ha G, Yap D, Farahani H, et al. Dynamics of genomic clones in breast cancer patient xenografts at single-cell resolution. Nature. 2014;.})
}
\keyword{datasets}

